67: Changed sensor value datatypes (and limits) to integers instead of floats

+++ access_point/database/database.py
@@ -74,7 +74,7 @@ class Database:
                         sensor_name: str,
                         unit: Optional[str],
                         timestamp: datetime,
-                        value: float,
+                        value: int,
                         alarm: Literal['n', 'l', 'h']) -> None:
         """
         Adds a measured value for a single sensor of a given sensor station to the database.
@@ -176,8 +176,8 @@ class Database:
     def update_sensor_setting(self,
                               sensor_station_address: str,
                               sensor_name: str,
-                              lower_limit: Optional[float] = None,
-                              upper_limit: Optional[float] = None,
+                              lower_limit: Optional[int] = None,
+                              upper_limit: Optional[int] = None,
                               alarm_tripping_time: Optional[int] = None,
                               **kwargs) -> None:
         """
@@ -242,7 +242,7 @@ class Database:
                 "sensor_name": Name of the sensor -> str,
                 "unit": Unit of the measured value -> str | None,
                 "timestamp": Timestamp of the measurement -> datetime,
-                "value": Measured value -> float,
+                "value": Measured value -> int,
                 "alarm": Alarm active at the time of the measurement -> str ['n' no alarm | 'l' below limit | 'h' above limit]
             }
         """
@@ -280,7 +280,7 @@ class Database:
     @_with_connection
     def get_limits(self,
                    sensor_station_address: str,
-                   sensor_name: str) -> tuple[Optional[float], Optional[float], Optional[timedelta], datetime]:
+                   sensor_name: str) -> tuple[Optional[int], Optional[int], Optional[timedelta], datetime]:
         """
         Gets the currently set limits and timing information on limits.
         :param sensor_station_address: Address of the sensor station
